backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-Bcrypt
Flask-JWT-Extended
dotenv
pytest
flake8
Flask-Cors
Flask-Mail